feat(ui): move enrollment to Admin page, clean up dashboard
Move pending agents from dashboard to /admin/enrollment page with
accept/reject. Dashboard shows only active agents (pending filtered
out). Add Enrollment to Admin navbar dropdown. Add skipped to job
status filter. Fix pending/running color semantics.
