+++ ui/src/pages/dashboard.tsx
@@ -1,9 +1,7 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useMemo } from "react";
 import { useVimScroll } from "@/hooks/use-vim-scroll";
-import { useCommands } from "@/hooks/use-commands";
 import { ContentArea } from "@/components/layout/content-area";
 import { AgentCard } from "@/components/domain/agent-card";
-import { PendingAgentCard } from "@/components/domain/pending-agent-card";
 import { ComponentRow } from "@/components/domain/component-row";
 import { Card, CardContent } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
@@ -15,14 +13,8 @@ import { DataTable } from "@/components/ui/data-table";
 import { Text } from "@/components/ui/text";
 import { useHealth } from "@/hooks/use-health";
 import { useAgents } from "@/hooks/use-agents";
-import { useAuth } from "@/hooks/use-auth";
-import { Activity, Loader2, ShieldAlert } from "lucide-react";
-import type { ComponentHealth, ComponentEntry, PendingAgentInfo } from "@/sdk/gen/schemas";
-import {
-  getAgentsPending,
-  acceptAgent,
-  rejectAgent,
-} from "@/sdk/gen/agent-management-api-agent-operations/agent-management-api-agent-operations";
+import { Activity, Loader2 } from "lucide-react";
+import type { ComponentHealth, ComponentEntry } from "@/sdk/gen/schemas";
 
 function formatBytes(bytes: number) {
   if (bytes < 1024) return `${bytes} B`;
@@ -101,81 +93,11 @@ export function Dashboard() {
     loading: agentLoading,
     refresh: refreshAgents,
   } = useAgents();
-  const { can } = useAuth();
 
-  const [pendingAgents, setPendingAgents] = useState<PendingAgentInfo[]>([]);
-
-  const fetchPending = useCallback(async () => {
-    if (!can("agent:write")) return;
-    try {
-      const resp = await getAgentsPending();
-      if (resp.status === 200) {
-        setPendingAgents(
-          (resp.data as { agents?: PendingAgentInfo[] })?.agents ?? [],
-        );
-      }
-    } catch {
-      // silent — pending section just won't show
-    }
-  }, [can]);
-
-  useEffect(() => {
-    fetchPending();
-    const interval = setInterval(fetchPending, 10000);
-    return () => clearInterval(interval);
-  }, [fetchPending]);
-
-  const refreshAll = useCallback(() => {
-    refreshAgents();
-    fetchPending();
-  }, [refreshAgents, fetchPending]);
-
-  // Command bar commands for PKI operations.
-  useCommands(
-    [
-      {
-        id: "pki:pending",
-        name: "pending",
-        description: "Show pending PKI enrollments",
-        category: "navigation",
-        action: () => {
-          document.getElementById("pending-agents")?.scrollIntoView({ behavior: "smooth" });
-        },
-      },
-      ...pendingAgents.map((a) => ({
-        id: `pki:accept:${a.hostname}`,
-        name: `accept ${a.hostname}`,
-        description: `Accept PKI enrollment for ${a.hostname}`,
-        category: "actions",
-        action: async () => {
-          await acceptAgent(a.hostname);
-          fetchPending();
-          refreshAgents();
-        },
-      })),
-      ...pendingAgents.map((a) => ({
-        id: `pki:reject:${a.hostname}`,
-        name: `reject ${a.hostname}`,
-        description: `Reject PKI enrollment for ${a.hostname}`,
-        category: "actions",
-        action: async () => {
-          await rejectAgent(a.hostname);
-          fetchPending();
-        },
-      })),
-    ],
-    [pendingAgents, fetchPending, refreshAgents],
-  );
-
-  // Filter out pending agents from the main agent list — they're
-  // shown in the Pending Enrollment section with accept/reject actions.
-  const pendingMachineIds = useMemo(
-    () => new Set(pendingAgents.map((pa) => pa.machine_id)),
-    [pendingAgents],
-  );
+  // Filter out pending agents — they're managed in Admin > Enrollment.
   const activeAgents = useMemo(
-    () => agents.filter((a) => !pendingMachineIds.has(a.machine_id ?? "")),
-    [agents, pendingMachineIds],
+    () => agents.filter((a) => a.state !== "Pending"),
+    [agents],
   );
 
   const loading = healthLoading || agentLoading;
@@ -397,26 +319,8 @@ export function Dashboard() {
             </div>
           )}
 
-          {/* Pending Agents (PKI enrollment) */}
-          {pendingAgents.length > 0 && can("agent:write") && (
-            <section id="pending-agents" className="mt-6">
-              <SectionLabel icon={ShieldAlert}>
-                Pending Enrollment ({pendingAgents.length})
-              </SectionLabel>
-              <div className="grid gap-3 sm:grid-cols-2 lg:grid-cols-3">
-                {pendingAgents.map((pa) => (
-                  <PendingAgentCard
-                    key={pa.machine_id}
-                    agent={pa}
-                    onRefresh={refreshAll}
-                  />
-                ))}
-              </div>
-            </section>
-          )}
-
-          {/* Agents — exclude pending agents (shown above) */}
-          <section className="mt-6">
+          {/* Agents — pending agents are managed in Admin > Enrollment */}
+          <section>
             <SectionLabel icon={Activity}>
               Agents ({activeAgents.length})
             </SectionLabel>
@@ -426,7 +330,7 @@ export function Dashboard() {
                   key={agent.hostname}
                   agent={agent}
                   components={grouped?.agent}
-                  onRefresh={refreshAll}
+                  onRefresh={refreshAgents}
                 />
               ))}
             </div>
